had to remove all the orderbook code. too many issues. redid https client so it runs in its own thread instead of each widget creating its own version. also transitioning this to a pub-sub type architecture.

+++ common/include/BookUpdate.hpp
@@ -15,8 +15,8 @@ struct EntryUpdate {
 struct BookUpdate {
   std::string symbol;
 
-  unsigned int first_update;
-  unsigned int last_update;
+  unsigned long first_update;
+  unsigned long last_update;
 
   std::vector<EntryUpdate> bids;
   std::vector<EntryUpdate> asks;
